requirements.txt
requests
beautifulsoup4
psycopg2-binary
fastapi
uvicorn
sqlalchemy
google-generativeai
python-dotenv
python-multipart
